pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "Bishe_LLM_for_genesis"
version = "0.1.0"
description = "LLM fine-tuning, evaluation, and simulation control pipeline for Genesis."
readme = "PROJECT_STRUCTURE.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Bishe Team" }
]
dependencies = [
  "typer>=0.12.0",
  "pyyaml>=6.0.1",
  "torch>=2.2.0",
  "transformers>=4.40.0",
  "datasets>=2.18.0",
  "accelerate>=0.28.0",
  "bitsandbytes>=0.43.0",
  "vllm>=0.4.0"
]

[project.scripts]
genesis-cli = "cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "genesis_example*"]

[tool.setuptools.package-data]
"src.data" = ["configs/*.json"]
